Organized the test directory into folders
