+++ test/mocha.opts
@@ -1,2 +1,3 @@
 --compilers js:babel-register
 --require babel-polyfill
+--recursive
