+++ test/routes-tests/projects.routes.test.js
@@ -1,12 +1,12 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
 
-import {defineTables, populateTools} from '../src/db/init_database';
-import { populateUsers, populateProjects } from '../src/db/import_test_data';
+import {defineTables, populateTools} from '../../src/db/init_database';
+import { populateUsers, populateProjects } from '../../src/db/import_test_data';
 // Using Expect style
-const sequelize = require('../src/db/sequelize');
-import * as apiUtil from './util/api.util';
-import server from '../src';
+const sequelize = require('../../src/db/sequelize');
+import * as apiUtil from '../util/api.util';
+import server from '../../src';
 
 chai.use(chaiHttp);
 const should = chai.should();
