+++ test/service-tests/users.service.test.js
@@ -1,11 +1,11 @@
 import { expect } from 'chai';
 
-import { defineTables } from '../src/db/init_database';
-import { populateUsers } from '../src/db/import_test_data';
+import { defineTables } from '../../src/db/init_database';
+import { populateUsers } from '../../src/db/import_test_data';
 // Using Expect style
-const sequelize = require('../src/db/sequelize');
-import UserService from './../src/services/users.service';
-import { getMockLogger } from './mockLogger';
+const sequelize = require('../../src/db/sequelize');
+import UserService from '../../src/services/users.service';
+import { getMockLogger } from '../util/mockLogger';
 
 sequelize.initSequelize();
